Dashboard

+++ backend/users-service/src/main/java/org/meristem/oneapp/usersservice/domains/responses/SendOtpResponse.java
@@ -3,5 +3,5 @@ package org.meristem.oneapp.usersservice.domains.responses;
 import lombok.Builder;
 
 @Builder
-public record SendOtpResponse(String message, String recipient, Short timeToExpireInSeconds) {
+public record SendOtpResponse(String message, String recipient, Integer timeToExpireInSeconds) {
 }
